+++ pkgs/native_assets_builder/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## 0.2.0-wip
 
-- **Breaking change** `NativeAssetsBuildRunner`s method now return an object
+- **Breaking change** `NativeAssetsBuildRunner`s methods now return an object
   ([#105](https://github.com/dart-lang/native/issues/105)).
+- **Breaking change** `NativeAssetsBuildRunner`s methods now return value now
+  contains a list of errors instead of it throwing
+  ([#106](https://github.com/dart-lang/native/issues/106)).
 - Use an `out/` sub directory for building native assets
   ([#98](https://github.com/dart-lang/native/issues/98)).
 - Check asset ids on having having a package uri with the owning package